Moved Impl::Hull to quickhull.cpp
