+++ src/manifold/src/impl.cpp
@@ -478,29 +478,6 @@ Manifold::Impl::Impl(const Mesh& mesh, const MeshRelationD& relation,
   Finish();
 }
 
-void Manifold::Impl::Hull(const std::vector<glm::vec3>& vertPos) {
-  size_t numVert = vertPos.size();
-  if (numVert < 4) {
-    status_ = Error::InvalidConstruction;
-    return;
-  }
-
-  Vec<glm::dvec3> pointCloudVec(numVert);
-  manifold::transform(vertPos.begin(), vertPos.end(), pointCloudVec.begin(),
-                      [](const glm::vec3& v) { return glm::dvec3(v); });
-  QuickHull qh(pointCloudVec);
-  std::tie(halfedge_, vertPos_) = qh.buildMesh();
-  meshRelation_.originalID = ReserveIDs(1);
-  CalculateBBox();
-  SetPrecision(bBox_.Scale() * kTolerance);
-  SplitPinchedVerts();
-  CalculateNormals();
-  InitializeOriginal();
-  CreateFaces({});
-  SimplifyTopology();
-  Finish();
-}
-
 /**
  * Create either a unit tetrahedron, cube or octahedron. The cube is in the
  * first octant, while the others are symmetric about the origin.
